turn base fixtures into hooks
